Merging the dev work from S2 into master

+++ server/controllers/admin/verifyUser.js
@@ -1,16 +1,25 @@
 /**
  * update the user state to be verified
  * remove the verificcation image from DB
- * @todo delete verification image from google storage
  * @param {id} mongoID - user id
  */
 
 const boom = require("boom");
+const approvalEmail = require("./../../helpers/emails/approvalEmail");
 
-const { updateUserById, getUserById } = require("./../../database/queries/user");
+
+const {
+  updateUserById,
+  getUserById,
+  updateUserHelpfulPoints,
+} = require("./../../database/queries/user");
+const { updateHelpfulPoints } = require("./../../database/queries/reviews");
+
 const deleteFile = require("./../../helpers/deleteFile");
 
-module.exports = ((req, res, next) => {
+const { referralPoints } = require("./../../constants");
+
+module.exports = async (req, res, next) => {
   const { id } = req.body;
   const updateData = {
     awaitingReview: false,
@@ -18,32 +27,48 @@ module.exports = ((req, res, next) => {
     verificationPhoto: undefined,
   };
 
-  getUserById(id)
-    .then((user) => {
-      if (!user) {
-        next(boom.notFound("user not found!"));
-      } else if (!user.verificationPhoto) {
-        next(boom.badData("the user has no verification image"));
-      } else {
-        // update user state
-        updateUserById(id, updateData)
-          .then(() => {
-            // delete verification photo from google storage
-            deleteFile(user.verificationPhoto)
-              .then(() => {
-                res.send();
-              })
-              .catch((err) => {
-                if (err.message === "file is no longer available") {
-                  next(boom.badData(err));
-                } else {
-                  next(boom.badImplementation());
-                }
-              });
-          })
-          .catch(() => {
-            next(boom.badImplementation());
-          });
+  try {
+    const user = await getUserById(id);
+
+    if (!user) {
+      return next(boom.notFound("user not found!"));
+    }
+    if (!user.verificationPhoto) {
+      return next(boom.badData("the user has no verification image"));
+    }
+    // update user state
+    await updateUserById(id, updateData);
+
+    try {
+      // delete verification photo from google storage
+      if (process.env.NODE_ENV !== "test") {
+        await deleteFile(user.verificationPhoto);
       }
-    });
-});
+    } catch (error) {
+      if (error.message === "file is no longer available") {
+        return next(boom.badData(error));
+      }
+      return next(boom.badImplementation());
+    }
+
+    if (user.referral) {
+      await updateHelpfulPoints({
+        points: referralPoints,
+        helpfulUser: user.referral,
+        helpedUser: id,
+        fromReferral: true,
+      });
+      await updateUserHelpfulPoints(user.referral);
+    }
+
+
+    if (process.env.NODE_ENV !== "test") {
+      // send aprroval email
+      await approvalEmail(user.email);
+    }
+
+    return res.json({});
+  } catch (error) {
+    return next(boom.badImplementation());
+  }
+};
